cli: store boundary, ceiling and bloom-bits as uint64_t
These three fastidious options are validated to fixed ranges (boundary
>= 2, ceiling in 40..2^30, bloom-bits in 2..64) but were stored as
int64_t and re-cast to uint64_t at their consumers (mass comparisons in
algod1_statistics.cc / algod1_fastidious.cc, the ceiling*one_megabyte
memory arithmetic, the bloom-bits asserts). Store them as uint64_t.

Their range checks move to free functions validate_boundary,
validate_ceiling and validate_bloom_bits that check the signed value
returned by args_long() (rejecting negatives before they could wrap)
and return the narrowed uint64_t; validate_fastidious keeps only the
cross-option checks (fastidious requires d=1; -b/-c/-y require -f). The
error messages are unchanged. The consumer casts are dropped (the
ceiling*one_megabyte product stays 64-bit and overflow-free since
one_megabyte is unsigned int), and the boundary log specifier becomes
PRIu64.

test_options.sh: 470 PASS, 0 FAIL.

+++ src/utils/algod1_fastidious.cc
@@ -359,7 +359,7 @@ namespace {
         ++heavy_state.amplicon;
         auto const & target_amplicon = ampinfo_v[heavy_amplicon_id];
         auto const & target_swarm = swarminfo_v[target_amplicon.swarmid];
-        if (target_swarm.mass >= static_cast<uint64_t>(parameters.opt_boundary))
+        if (target_swarm.mass >= parameters.opt_boundary)
           {
             ++heavy_state.progress;
             progress.update(heavy_state.progress);
@@ -428,7 +428,7 @@ namespace {
         --state.amplicon;
         auto const & target_amplicon = ampinfo_v[light_amplicon_id];
         auto const & target_swarm = swarminfo_v[target_amplicon.swarmid];
-        if (target_swarm.mass < static_cast<uint64_t>(parameters.opt_boundary))
+        if (target_swarm.mass < parameters.opt_boundary)
           {
             ++state.progress;
             progress.update(state.progress);
